Update styles.css
Since textarea and div with marks overlap each other, sometimes focus could be on one or another. When focus is no textarea and it is scrolled, we programatically scroll backdrop div with marks by `handleScroll` method. However, when backdrop is in focus and scrolled, we don't do vice versa, so backdrop with marks is scrolled, while textearea stays unscrolled, so they misaligned. The easiest way to handle it is to add `pointer-events: none;` to backdrop styles, so it skips all clicks\scrolls and other events and becomes "invisible" allowing textarea to catch it and call same `handleScroll`